chore: split file-table in react components
